feature[test]: refactoring issue #389
